requirements.txt
freetype-py==2.3.0
Pillow==9.5.0
pycairo==1.23.0
PyYAML==6.0
reportlab==4.0.0
rlPyCairo==0.2.0
